latest local comitt.

+++ src/test/java/com/tracebucket/crm/test/repository/PartnerRepositoryTest.java
@@ -24,12 +24,12 @@ import javax.transaction.Transactional;
  * Created by vishwa on 13-01-2015.
  */
 
-/*@RunWith(SpringJUnit4ClassRunner.class)
+@RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes  = {ApplicationTestConfig.class, JPATestConfig.class})
-@Transactional*/
+@Transactional
 public class PartnerRepositoryTest {
 
-/*    @Autowired
+   @Autowired
     private PartnerRepository partnerRepository;
 
     @Autowired
@@ -44,13 +44,13 @@ public class PartnerRepositoryTest {
 
     }
 
-  /*  private void createPartner() throws Exception{
-        partner = PartnerFixture.allRoles();
-        organization = organizationRepository.save(partner.getOwner());
+    private void createPartner() throws Exception{
+        partner = PartnerFixture.standardPartner();
+        //organization = organizationRepository.save(partner.getOwner());
         partner = partnerRepository.save(partner);
     }
 
-    @Test
+  /*  @Test
     public void testUpdate() throws Exception {
         createPartner();
         partner.getPartnerRoles().clear();
@@ -60,18 +60,17 @@ public class PartnerRepositoryTest {
         partner = partnerRepository.save(partner);
         Assert.assertEquals(1, partner.getPartnerRoles().size());
     }
-*/
-  /*  @Test
+    @Test
     @Rollback(false)
     public void testFindById() throws Exception {
         createPartner();
         partner = partnerRepository.findOne(partner.getAggregateId());
         Assert.assertNotNull(partner);
-    }*/
+    }
 
     @After
     public void tearDown(){
-/*        if(partner != null && partner.getAggregateId() != null) {
+        if(partner != null && partner.getAggregateId() != null) {
             partnerRepository.delete(partner.getAggregateId());
             partner = partnerRepository.findOne(partner.getAggregateId());
             Assert.assertNull(partner);
@@ -80,14 +79,14 @@ public class PartnerRepositoryTest {
             organizationRepository.delete(organization.getAggregateId());
             organization = organizationRepository.findOne(organization.getAggregateId());
             Assert.assertNull(organization);
-        }*/
-    }
+        }
+    }*/
 
-   /* @Test
+    @Test
     @Rollback(false)
     public void testCreate() throws Exception{
         createPartner();
         Assert.assertNotNull(partner.getAggregateId());
-    }*/
+    }
 
 }
